app_nnpol: fp16 weights in the slot - the 64x64x64 family fits the 28 KB of SRAM the image has spare
The 48 KB fp32 slot overflowed RAM by 15 KB (34 KB spare before it),
and no fp32 3x64 blob is under 41 KB. The slot now carries IEEE fp16
weights: 20-26 KB for the family, in a 28 KB .bss buffer. The
interpreter converts each weight as it reads it - on the target GCC's
__fp16 (the firmware builds with -mfp16-format=ieee) is one vcvtb after
the halfword load, on the host parity build a software conversion; both
are exact, so the two agree bit for bit. The blob magic is "NPS2"
(payload floats are 2 bytes, padded to a word), the validator's size
rule follows, everything else in the format is unchanged.

The exporter (mjx-drone-trainer export_policy_c.py) writes the fp16
payload and computes policy_reference.npz with the rounded weights, so
the harness's network check stays exact. The host shadow keeps the fp32
checkpoint and differs from the vehicle by the rounding, ~1e-3 in the
action.

+++ examples/app_nnpol/src/nnpol.h
@@ -96,7 +96,7 @@ typedef struct {
 /** A validated slot: its header and the weights behind it. */
 typedef struct {
   const nnpolSlotHeader_t* hdr;
-  const float* weights;
+  const uint16_t* weights;   /**< fp16, see nnpol_slot.h */
 } nnpolPolicy_t;
 
 /** Reference position at time t (seconds), world frame, BEFORE the yaw
